requirements.txt
streamlit
torch
transformers
Pillow
gdown
huggingface-hub
numpy
tqdm
pandas
requests
